disable openId in production using a config option. This lets us figure out how to we want to do key storage in production
