+++ backend/LexBoxApi/Auth/OpenIdOptions.cs
@@ -1,0 +1,9 @@
+﻿using System.ComponentModel.DataAnnotations;
+
+namespace LexBoxApi.Auth;
+
+public class OpenIdOptions
+{
+    [Required]
+    public required bool Enable { get; set; }
+}
