+++ backend/LexBoxApi/appsettings.Development.json
@@ -7,10 +7,10 @@
     }
   },
   "DbConfig": {
-    "LexBoxConnectionString": "Host=localhost;Port=5433;Username=postgres;Password=972b722e63f549938d07bd8c4ee5086c;Database=lexbox;Include Error Detail=true",
+    "LexBoxConnectionString": "Host=localhost;Port=5433;Username=postgres;Password=972b722e63f549938d07bd8c4ee5086c;Database=lexbox;Include Error Detail=true"
   },
   "LfClassicConfig": {
-    "ConnectionString": "mongodb://localhost:27017",
+    "ConnectionString": "mongodb://localhost:27017"
   },
   "ForwardedHeadersOptions": {
     "KnownNetworks": [
@@ -54,6 +54,9 @@
     "Google": {
       "ClientId": "694338503380-rf3t82f3kqc7l0orjletk2r58lep0lvi.apps.googleusercontent.com",
       "ClientSecret": "__REPLACE__"
+    },
+    "OpenId": {
+      "Enable": true
     }
   },
   "Tus" : {
