+++ backend/LexBoxApi/appsettings.json
@@ -60,6 +60,9 @@
 //      properties declared in CookieAuthenticationOptions
 //      15 days, 0 hours
       "ExpireTimeSpan": "15.00:00"
+    },
+    "OpenId": {
+      "Enable": false
     }
   },
   "Otel": {
